Quality Notes
- functional
  - smoke test on demo page and e2e; change filters to cause server data fetch and confirm data updates
- perf (50k users in one course)
  - before: on load, ~750ms various processing, plus ~450ms in `onServerDataReload`
  - after: on load, ~275ms various processing; ~10ms in `onServerDataReload`
- devops/security/ux: no change